+++ src/algo/mod.rs
@@ -14,6 +14,7 @@ pub mod dijkstra_dist;
 pub mod dijkstra_pred;
 pub mod distance_matrix;
 pub mod floyd_warshall;
+pub mod johnson_75;
 pub mod predecessor_tree;
 pub mod tarjan;