+++ src/repr/adjacency_list_weighted/mod.rs
@@ -57,15 +57,15 @@ impl<W> AddArcWeighted<W> for AdjacencyListWeighted<W> {
     /// # Panics
     ///
     /// * Panics if `u` equals `v`.
-    /// * Panics if `u` is out of bounds.
-    /// * Panics if `v` is out of bounds.
+    /// * Panics if `u` is not in the digraph.
+    /// * Panics if `v` is not in the digraph.
     fn add_arc_weighted(&mut self, u: usize, v: usize, w: W) {
         assert_ne!(u, v, "u = {u} equals v = {v}");
 
         let order = self.order();
 
-        assert!(u < order, "u = {u} is out of bounds");
-        assert!(v < order, "v = {v} is out of bounds");
+        assert!(u < order, "u = {u} is not in the digraph");
+        assert!(v < order, "v = {v} is not in the digraph");
 
         let _ = self.arcs[u].insert(v, w);
     }
@@ -180,17 +180,22 @@ impl From<AdjacencyMatrix> for AdjacencyListWeighted<usize> {
     }
 }
 
-impl<W> From<Vec<BTreeMap<usize, W>>> for AdjacencyListWeighted<W> {
+impl<W, I> From<I> for AdjacencyListWeighted<W>
+where
+    I: IntoIterator<Item = BTreeMap<usize, W>>,
+{
     /// # Panics
     ///
     /// * Panics if, for any arc `u -> v` in `arcs`, `u` equals `v`.
-    /// * Panics if, for any arc `u -> v` in `arcs`, `v` is out of bounds.
-    fn from(arcs: Vec<BTreeMap<usize, W>>) -> Self {
-        let digraph = Self { arcs };
+    /// * Panics if, for any arc `u -> v` in `arcs`, `v` is not in the digraph.
+    fn from(iter: I) -> Self {
+        let digraph = Self {
+            arcs: iter.into_iter().collect(),
+        };
 
         for (u, v) in digraph.arcs() {
             assert_ne!(u, v, "u = {u} equals v = {v}");
-            assert!(v < digraph.order(), "v = {v} is out of bounds");
+            assert!(v < digraph.order(), "v = {v} is not in the digraph");
         }
 
         digraph
@@ -206,9 +211,9 @@ impl<W> HasArc for AdjacencyListWeighted<W> {
 impl<W> Indegree for AdjacencyListWeighted<W> {
     /// # Panics
     ///
-    /// Panics if `v` is out of bounds.
+    /// Panics if `v` is not in the digraph.
     fn indegree(&self, v: usize) -> usize {
-        assert!(v < self.order(), "v = {v} is out of bounds");
+        assert!(v < self.order(), "v = {v} is not in the digraph");
 
         self.arcs
             .iter()
@@ -229,7 +234,7 @@ impl<W> IsSimple for AdjacencyListWeighted<W> {
 impl<W> OutNeighbors for AdjacencyListWeighted<W> {
     /// # Panics
     ///
-    /// Panics if `u` is out of bounds.
+    /// Panics if `u` is not in the digraph.
     fn out_neighbors(&self, u: usize) -> impl Iterator<Item = usize> {
         self.arcs[u].keys().copied()
     }
@@ -238,7 +243,7 @@ impl<W> OutNeighbors for AdjacencyListWeighted<W> {
 impl<W> OutNeighborsWeighted<W> for AdjacencyListWeighted<W> {
     /// # Panics
     ///
-    /// Panics if `u` is out of bounds.
+    /// Panics if `u` is not in the digraph.
     fn out_neighbors_weighted<'a>(
         &'a self,
         u: usize,
@@ -259,7 +264,7 @@ impl<W> Order for AdjacencyListWeighted<W> {
 impl<W> Outdegree for AdjacencyListWeighted<W> {
     /// # Panics
     ///
-    /// Panics if `u` is out of bounds.
+    /// Panics if `u` is not in the digraph.
     fn outdegree(&self, u: usize) -> usize {
         self.arcs[u].len()
     }
@@ -581,13 +586,13 @@ mod tests {
     }
 
     #[test]
-    #[should_panic(expected = "v = 1 is out of bounds")]
+    #[should_panic(expected = "v = 1 is not in the digraph")]
     fn add_arc_weighted_out_of_bounds_u() {
         AdjacencyListWeighted::trivial().add_arc_weighted(0, 1, 1);
     }
 
     #[test]
-    #[should_panic(expected = "u = 1 is out of bounds")]
+    #[should_panic(expected = "u = 1 is not in the digraph")]
     fn add_arc_weighted_out_of_bounds_v() {
         AdjacencyListWeighted::trivial().add_arc_weighted(1, 0, 1);
     }
@@ -1334,7 +1339,7 @@ mod tests {
     }
 
     #[test]
-    #[should_panic(expected = "v = 1 is out of bounds")]
+    #[should_panic(expected = "v = 1 is not in the digraph")]
     fn from_vec_out_of_bounds_v() {
         let vec = vec![BTreeMap::from([(1, -1)])];
         let _ = AdjacencyListWeighted::from(vec);
@@ -1536,7 +1541,7 @@ mod tests {
     }
 
     #[test]
-    #[should_panic(expected = "v = 1 is out of bounds")]
+    #[should_panic(expected = "v = 1 is not in the digraph")]
     fn indegree_out_of_bounds() {
         let _ = AdjacencyListWeighted::<usize>::trivial().indegree(1);
     }