+++ src/repr/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod adjacency_list;
 pub mod adjacency_list_weighted;
+pub mod adjacency_map;
 pub mod adjacency_matrix;
 pub mod edge_list;
 pub mod test_unweighted;
@@ -9,6 +10,7 @@ pub mod test_unweighted;
 pub use {
     adjacency_list::AdjacencyList,
     adjacency_list_weighted::AdjacencyListWeighted,
+    adjacency_map::AdjacencyMap,
     adjacency_matrix::AdjacencyMatrix,
     edge_list::EdgeList,
 };